Refactor unified turn module into submodules
